% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderSelectionDlg.R
\name{.orderSelectionDlg}
\alias{.orderSelectionDlg}
\title{Select ordering mode}
\usage{
.orderSelectionDlg()
}
\value{
Returns one of:

- `"asc"`
- `"desc"`
- `"rand"`

Returns `NULL` if cancelled.
}
\description{
Small Tcl/Tk dialog to choose how selected code elements
should be ordered.
}
