% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggleUnicode.R
\name{escapeUnicode}
\alias{escapeUnicode}
\title{Escape Unicode Characters}
\usage{
escapeUnicode(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector containing Unicode escape sequences
}
\description{
Converts non-ASCII characters to Unicode escape sequences.
}
\examples{
escapeUnicode("Schneeh\u00f6he")
unescapeUnicode("Schneeh\\\\u00f6he")

}
