% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flushToSource.R
\name{flushToSource}
\alias{flushToSource}
\title{Insert selected object as reproducible R code}
\usage{
flushToSource(width.cutoff = 500L)
}
\arguments{
\item{width.cutoff}{Integer. Passed to [base::deparse()].}
}
\value{
Invisibly returns the generated code.
}
\description{
Converts the currently selected object in the RStudio editor into
reproducible R source code and inserts it into the active document.
}
\details{
Data frames are converted into a readable `data.frame(...)`
representation instead of the more verbose `dput()` structure output.

Other objects are serialized using [base::dput()].
}
\examples{
\dontrun{
flushToSource()
}

}
