% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some.R
\name{some}
\alias{some}
\alias{some.data.frame}
\alias{some.matrix}
\alias{some.default}
\title{Random subset of an object}
\usage{
some(x, n = 6L, ...)

\method{some}{data.frame}(x, n = 6L, ...)

\method{some}{matrix}(x, n = 6L, addrownums = TRUE, ...)

\method{some}{default}(x, n = 6L, ...)
}
\arguments{
\item{x}{An R object.}

\item{n}{Integer scalar specifying the number of elements or rows
to return. Defaults to `6L`.
Negative values reduce the total number of returned elements
relative to the object size.}

\item{...}{Additional arguments passed to methods.}

\item{addrownums}{Logical. Should artificial row names be added
for matrices without row names? Only used for matrices.}
}
\value{
A random subset of `x`.

- For vectors: a vector.
- For matrices: a matrix.
- For data frames: a data frame.
}
\description{
Returns a random subset of observations/elements from vectors,
matrices, and data frames.
}
\details{
Negative values of `n` are interpreted as:

`length(x) + n`

or

`nrow(x) + n`

respectively.

This function is useful for quickly inspecting random parts of
larger objects.
}
\examples{
some(1:100)

some(letters, 10)

some(mtcars, 5)

some(as.matrix(mtcars), 4)

# Negative n
some(1:20, -5)

}
