% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggleUnicode.R
\name{unescapeUnicode}
\alias{unescapeUnicode}
\title{Unescape Unicode Characters}
\usage{
unescapeUnicode(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector containing decoded Unicode characters
}
\description{
Converts Unicode escape sequences to their corresponding characters.
}
\examples{
unescapeUnicode("Schneeh\\\\u00f6he")

}
