% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggleUnicode.R
\name{unicodeAddins}
\alias{unicodeAddins}
\alias{toggleUnicodeAddin}
\title{Unicode RStudio Addins}
\usage{
toggleUnicodeAddin()
}
\value{
`NULL`, invisibly
}
\description{
Converts Unicode characters or escape sequences in the selected text.
}
